VideoToQ&A/requirements.txt
pyautogui
opencv-python
pytesseract
python-docx
numpy
Pillow
